Combined ExpressionSerializer.java and ExpressionDeserializer.java into single class (ExpressionComponent.java). Added @Lazy annotation to WeaponFactory. Updated BonusDamageService to reflect record changes to Receiver.
