+++ src/main/java/Tekiz/_DPSCalculator/_DPSCalculator/model/perks/Perk.java
@@ -3,8 +3,7 @@ package Tekiz._DPSCalculator._DPSCalculator.model.perks;
 import Tekiz._DPSCalculator._DPSCalculator.model.enums.modifiers.ModifierSource;
 import Tekiz._DPSCalculator._DPSCalculator.model.enums.modifiers.ModifierTypes;
 import Tekiz._DPSCalculator._DPSCalculator.model.modifiers.Modifier;
-import Tekiz._DPSCalculator._DPSCalculator.util.deserializer.ExpressionDeserializer;
-import Tekiz._DPSCalculator._DPSCalculator.util.deserializer.ExpressionSerializer;
+import Tekiz._DPSCalculator._DPSCalculator.util.deserializer.ExpressionComponent.*;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
